made larger components into smaller components
