+++ src/services/auth/SupabaseAuth.ts
@@ -597,7 +597,6 @@ export class SupabaseAuth implements ServiceAuth {
           reward: null,
           stars: null,
           is_wa_contact: null,
-          tc_agreed_version: 0,
         });
         this._currentUser = createdUser;
       }
@@ -721,7 +720,6 @@ export class SupabaseAuth implements ServiceAuth {
           reward: null,
           stars: null,
           is_wa_contact: null,
-          tc_agreed_version: 0,
         });
       }
